requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
easyocr==1.7.0
yolov5==7.0.13
# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-mock==3.12.0

# Development utilities
black==23.12.0
flake8==6.1.0
mypy==1.7.1